fix: widen all money fields from int to long to prevent overflow on prices >= Rp 21.5M
SQLite INTEGER is already 64-bit; no schema migration needed. Also widens
CalculateAverageCost return type (was silently truncating via (int) cast),
PricingEngine.GetUnitPrice return type, PriceChangeEntry fields, and all
affected local variables, casts, and repository GetInt→GetLong calls.

+++ Kasir.Avalonia/Forms/Master/PriceChangeView.axaml.cs
@@ -120,7 +120,7 @@ public partial class PriceChangeView : UserControl
         foreach (var row in _allRows)
         {
             if (!decimal.TryParse(row.NewPrice, out decimal newVal)) continue;
-            int newPrice = (int)(newVal * 100m);
+            long newPrice = (long)(newVal * 100m);
             if (newPrice == row.Tag.Price) continue;
             changes.Add(new PriceChangeEntry { ProductCode = row.Tag.ProductCode, NewPrice = newPrice });
         }
