+++ Kasir.Core/Data/Repositories/PurchaseRepository.cs
@@ -136,7 +136,7 @@ namespace Kasir.Data.Repositories
                 Remark = SqlHelper.GetString(r, "remark"),
                 Quantity = SqlHelper.GetInt(r, "quantity"),
                 Value = SqlHelper.GetLong(r, "value"),
-                UnitPrice = SqlHelper.GetInt(r, "unit_price"),
+                UnitPrice = SqlHelper.GetLong(r, "unit_price"),
                 DiscPct = SqlHelper.GetInt(r, "disc_pct"),
                 DiscValue = SqlHelper.GetLong(r, "disc_value"),
                 Unit = SqlHelper.GetString(r, "unit")
