+++ Kasir.Core/Models/Product.cs
@@ -11,18 +11,18 @@ namespace Kasir.Models
         public string Status { get; set; }
         public string Unit { get; set; }
         public string Unit1 { get; set; }
-        public int Price { get; set; }
-        public int Price1 { get; set; }
-        public int Price2 { get; set; }
-        public int Price3 { get; set; }
-        public int Price4 { get; set; }
-        public int BuyingPrice { get; set; }
+        public long Price { get; set; }
+        public long Price1 { get; set; }
+        public long Price2 { get; set; }
+        public long Price3 { get; set; }
+        public long Price4 { get; set; }
+        public long BuyingPrice { get; set; }
         public string VendorCode { get; set; }
         public int QtyBreak2 { get; set; }
         public int QtyBreak3 { get; set; }
         public string OpenPrice { get; set; }
         public int DiscPct { get; set; }
-        public int CostPrice { get; set; }
+        public long CostPrice { get; set; }
         public string VatFlag { get; set; }
         public string LuxuryTaxFlag { get; set; }
         public string IsConsignment { get; set; }
